Implementation of Lex Programs
(a) General Algorithm/Pseudo code: the given questions are 
1.	Converting upper case letter to lowercase letters
2.	Count number of words
3.	Count words greater than 10 and truncate the rest of portion
4.	Add line number of lines to program
5.	Identify all numbers from the program
6.	Write program to count number of new lines or words that start with a
7.	Write program to add line number to a program
8.	Remove comments from source file
9.	WAP to catch all array declarations and multiple commas separated declarations. Send the caught declarations to new file.
10.	Choose a language of your choice and write lex program to generate its tokens. Built it as extensive as possible.

(b) Textual Description of the process of implementation
• We were given the sample exercise as above shown.
• We proceed with firstly installing the required environment from the internet. Which included flex, bison setups.
• Then we installed and set up these setups in a proper directory and begin with our coding.

c) Learning from the experiment:-
This experiment taught us a lot about the lexical phase itself. We got to know how the lexical files, i.e. .l files are made. We got to know the techniques for these exercises and tried to develop our own logic to build these up. Moreover, running a .exe file from the flex .l files was also a new thing added in our knowledge.