Coding for NFA for the given RE
a) General Algorithm / Pseudo Code
•	Input for the code: a valid regular expression
•	Output: the corresponding NFA generated for the given expression

b) Textual Description of the Process of Implementation
•	We were to design a code for converting the given regular expression into the NDA.
•	We proceed with the approach to counter the difficulty caused by infix assignments as inputted by user, thus we went on to take help of the open-source code to convert our infix input into a postfix one. Thus, we now get a prefix regular expression.
•	We then take care of the fact what data structure shall we use, thus we generate an NFA class where we define several functions like initial(), shownfa(), union(), concat(), closure() etc.
•	Our next step was implementing the given postfix operators using a stack[]. We kept on computing the values to and from this stack[].
•	Finally popping the last element gives us the desired NFA as such.

c) Learning from experiment :-
This experiment taught us that how to convert a regular expression into an NDA. This not only was an exciting mental problem but also was a great programming practice where we had to emphasis on the correct choice of data structures used. Moreover, getting the final output ready in a form of a graph is also very satisfying to watch that how the given re has transformed into a NFA with e moves.